refactor type resolver for java
