+++ java/fory-core/src/main/resources/META-INF/native-image/org.apache.fory/fory-core/native-image.properties
@@ -267,6 +267,7 @@ Args=--initialize-at-build-time=org.apache.fory.memory.MemoryBuffer,\
     org.apache.fory.reflect.TypeRef$TypeVariableKey,\
     org.apache.fory.reflect.TypeRef,\
     org.apache.fory.resolver.ClassChecker,\
+    org.apache.fory.resolver.TypeChecker,\
     org.apache.fory.resolver.ClassInfo,\
     org.apache.fory.resolver.ClassInfoHolder,\
     org.apache.fory.resolver.ClassResolver$2,\
